% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lambda.R
\name{setup_lambda}
\alias{setup_lambda}
\title{Generate lambda sequence.}
\usage{
setup_lambda(X, y, weights, lambda.min.ratio, nlambda)
}
\arguments{
\item{X}{Input matrix, of dimension \code{nobs} x \code{nvars}; 
each row is  an observation vector.}

\item{y}{Response variable, of length \code{nobs}. For \code{family="gaussian"}, 
it should be quantitative; for \code{family="binomial"}, it should be either 
a factor with two levels or a binary vector.}

\item{weights}{Observation weights.}

\item{lambda.min.ratio}{The smallest value for \code{lambda}, as a fraction of 
\code{lambda.max}, the smallest value for which all coefficients are zero. 
The default depends on the sample size \code{nobs} relative to the number 
of variables \code{nvars}.}

\item{nlambda}{The number of \code{lambda} values.}
}
\description{
Generate lambda sequence.
}
